Add test for limit=1, limit=Inf and default
