+++ packages/cli/test/fixtures/collect-static-dir-without-urls/board.html
@@ -1,0 +1,10 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <meta charset="utf-8" />
+    <title>board test page for staticDistDir usage</title>
+  </head>
+  <body>
+    test
+  </body>
+</html>
