+++ packages/cli/test/fixtures/collect-static-dir-without-urls/jobs.html
@@ -1,0 +1,10 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <meta charset="utf-8" />
+    <title>jobs test page for staticDistDir usage</title>
+  </head>
+  <body>
+    test
+  </body>
+</html>
