+++ packages/cli/test/fixtures/collect-static-dir-without-urls/team.html
@@ -1,0 +1,10 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <meta charset="utf-8" />
+    <title>team test page for staticDistDir usage</title>
+  </head>
+  <body>
+    test
+  </body>
+</html>
